+++ bootstrap-biz/src/main/resources/sql/bootstrap.h2.sql
@@ -1,5 +1,5 @@
 DROP TABLE IF EXISTS Greeting;
 CREATE TABLE `Greeting` (
-  `id` int PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint PRIMARY KEY AUTO_INCREMENT,
   `content` varchar(64) COMMENT 'content'
 );
